+++ components/PreviewPage.js
@@ -1,31 +1,16 @@
 /* eslint-disable react/no-unknown-property */
 import React from 'react';
-import { loadStripe } from '@stripe/stripe-js';
+
 import {connect} from 'react-redux'
 
 
-// Make sure to call `loadStripe` outside of a component’s render to avoid
-// recreating the `Stripe` object on every render.
-// eslint-disable-next-line no-unused-vars
-const stripePromise = loadStripe(
-  process.env.NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY
-);
+
 function PreviewPage({prova,boolean}) {
 
 
 
 
-  React.useEffect(() => {
-    // Check to see if this is a redirect back from Checkout
-    const query = new URLSearchParams(window.location.search);
-    if (query.get('success')) {
-     // console.log('Order placed! You will receive an email confirmation.');
-    }
 
-    if (query.get('canceled')) {
-      // console.log('Order canceled -- continue to shop around and checkout when you’re ready.');
-    }
-  }, []);
 
   return (
    // <form  action="/api/checkout_sessions" method="POST">
